refactor(selection): colocate samples limit alert
Inline the reactive limit alert and data attributes in the samples
table so the unused shared concern and partial can go away.
