--- app/views/shared/_selection_limit_alert.html.erb
@@ -1,13 +1,0 @@
-<%# locals: (message:, visible: false, alert_id: SelectionLimitableTable::SELECTION_LIMIT_ALERT_ID) %>
-
-<div
-  id="<%= alert_id %>"
-  class="<%= class_names("mb-4", hidden: !visible) %>"
-  role="alert"
-  aria-live="assertive"
-  data-selection-target="limitAlert"
->
-  <%= viral_alert(type: :warning, dismissible: false) do %>
-    <span data-selection-target="limitAlertMessage"><%= message %></span>
-  <% end %>
-</div>
